Add files via upload
The "Real-time Facial Expression Analysis and Quote Generation" project is a computer vision application that utilizes facial expression analysis to determine the dominant emotion expressed by an individual in real-time using their webcam. The project leverages the OpenCV and DeepFace libraries to capture frames from the camera and analyze the facial expressions.

The application continuously captures frames from the webcam and performs facial expression analysis on each frame. It identifies the dominant emotion expressed, such as happiness, sadness, anger, or surprise. The system keeps track of the most common expression observed over a predefined number of frames, providing an insight into the individual's prevailing emotion.

Furthermore, once the most common expression is determined, the project generates a random quote associated with that particular emotion. The quotes are stored in a dataset and retrieved based on the identified dominant emotion. This adds an interactive element to the application by providing users with inspiring or relatable quotes based on their expressed emotion.

The real-time analysis and quote generation are displayed on the screen alongside the video feed from the webcam. The dominant emotion and corresponding quote are overlaid as text on the video frames using the OpenCV library, enabling users to see their current mode and receive motivational or thought-provoking messages.

The "Real-time Facial Expression Analysis and Quote Generation" project can be used as a tool for emotional analysis, interactive experiences, or as a foundation for further research in areas like affective computing and human-computer interaction.